generator/konfig-docs/blog/2023/12/github-copilot-assets/slides/pyproject.toml
[tool.poetry]
name = "slides"
version = "0.1.0"
description = ""
authors = ["konfig-bot <bot@konfigthis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
opencv-python = "^4.8.1.78"
pyperclip = "^1.8.2"

[tool.poetry.group.dev.dependencies]
numpy = "^1.26.1"
Pillow = "^10.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
